move single_char_insert_string to its own module
